django/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
apturl==0.5.2
asgiref==3.7.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
autopep8==1.6.0
awscli==1.22.34
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
boto3==1.24.17
botocore==1.27.17
Brlapi==0.8.3
cachetools==4.2.2
celery==5.2.7
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
Deprecated==1.2.13
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==4.2.2
django-browser-reload==1.6.0
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
django-pdfkit==0.3.1
django-storages==1.12.3
django-tailwind==3.4.0
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djongo==1.3.6
docutils==0.17.1
duplicity==0.8.21
et-xmlfile==1.1.0
fasteners==0.14.1
filelock==3.6.0
frozenlist==1.3.1
future==0.18.2
GDAL==3.4.1
geoip2==4.6.0
google-api-core==2.0.1
google-api-python-client==2.22.0
google-auth==2.1.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-search-results==2.4.1
googleapis-common-protos==1.53.0
gunicorn==20.1.0
gyp==0.1
httplib2==0.19.1
idna==3.3
imgkit==1.2.2
importlib-metadata==4.6.4
iniconfig==1.1.1
jeepney==0.7.1
jmespath==1.0.1
keyring==23.5.0
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
maxminddb==2.2.0
model-bakery==1.6.0
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.2
netifaces==0.11.0
numpy==1.26.1
oauthlib==3.1.1
olefile==0.46
openpyxl==3.0.10
packaging==21.3
pandas==2.1.2
paramiko==2.9.3
pdfkit==1.0.0
pexpect==4.8.0
Pillow==9.2.0
platformdirs==2.5.1
pluggy==1.0.0
prompt-toolkit==3.0.31
protobuf==3.18.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycodestyle==2.8.0
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.0
PyJWT==2.4.0
pymacaroons==0.13.0
pymongo==3.12.3
PyNaCl==1.5.0
pyparsing==2.4.2
pyRFC3339==1.1
pytest==7.1.2
pytest-django==4.5.2
python-apt==2.3.0+ubuntu2
python-crontab==2.6.0
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.20.0
pytz==2022.2.1
pyxdg==0.27
PyYAML==5.4.1
random-username==1.0.2
redis==4.3.4
regex==2022.8.17
reportlab==3.6.8
requests==2.28.1
requests-oauthlib==1.3.0
roman==3.3
rsa==4.7.2
s3transfer==0.6.0
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
stripe==4.1.0
systemd-python==234
toml==0.10.2
tomli==2.0.1
typing_extensions==4.7.0
tzdata==2022.2
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
Unidecode==1.3.4
uritemplate==3.0.1
urllib3==1.26.12
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.5
wrapt==1.14.1
xdg==5
xkit==0.0.0
yarl==1.8.1
zipp==1.0.0
